updated with worm gear test using gpiozero library
