--- Spring2025/finger_sim/finger-simulation/requirements.txt
@@ -1,2 +1,0 @@
-Pygame
-numpy
