+++ Spring2025/motor_tests/src/motorcontrol10.py
