+++ Spring2025/motor_tests/src/motorcontrol2.py
