+++ Spring2025/motor_tests/src/run_motor.sh
